Fix two way binding for vms
According to Angular documentation (https://angular.io/guide/two-way-binding#how-two-way-binding-works),
the output decorator name for a "property" has to be "propertyChange".
To follow this pattern, the output decorator of "vms" must be called "vmsChange".

Moreover the EventEmitter's generic type variable needs to correspond to its property type.
Hence for vms of type {}[] the generic type of its EventEmitter must be {}[].