Use take(1) instead of first() to prevent EmptyErrors
If the observable does not emit any value and completes directly, using first() results in an error.

Using take(1) generally does the same as first() but with a minor difference.

take(1) mirrors the empty source observable in case it is empty and then also completes instead of throwing an error as first() does.